requirements.txt
sgiref==3.7.2
autopep8==2.0.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-crispy-forms==2.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4